Don't coerce types implicitly.

+++ core/src/com/unciv/logic/trade/TradeEvaluation.kt
@@ -62,7 +62,7 @@ class TradeEvaluation {
 
     fun getTradeAcceptability(trade: Trade, evaluator: Civilization, tradePartner: Civilization): Int {
         val citiesAskedToSurrender = trade.ourOffers.filter { it.type == TradeType.City }.count()
-        val maxCitiesToSurrender = ceil(evaluator.cities.size.toFloat() / 10)
+        val maxCitiesToSurrender = ceil(evaluator.cities.size.toFloat() / 10).toInt()
         if (citiesAskedToSurrender > maxCitiesToSurrender) {
             return Int.MIN_VALUE
         }
